Ajout de la pagination pour la liste des users

+++ Cours_MVC_1_Filezilla/app/view/commentaire/admin.php
@@ -25,18 +25,26 @@
 	<?php } ?>
 
 	<?php 
-		foreach($commentaires as $commentaire)
-		{ 
-			if($commentaire[0])
+
+		if(!$commentaires)
+		{ ?>
+			<div class="container bgcolor">
+				<p>Il n'y a aucun commentaire pour cet utilisateur pour le moment.</p>
+			</div>
+		<?php } 
+
+		else
+		{
+			foreach($commentaires as $commentaire)
 			{ ?>
-				<div class="container bgcolor">
-					<p>ID du commentaire: <?php echo $commentaire["comment_ID"]; ?></p>
-					<p>ID du post: <?php echo $commentaire["comment_post_ID"]; ?></p>
-					<p>ID auteur: <?php echo $commentaire["comment_author"]; ?></p>	
-					<p>Date: <?php echo $commentaire["comment_date"]; ?></p>
-					<p>Contenu: <?php echo $commentaire["comment_content"]; ?></p>
-					<a href="?module=commentaire&action=supprimer_commentaire&page=<?= $page ?>&id=<?= $commentaire['comment_ID']; ?>" onclick=" return confirm('Etes-vous certain de vous certain de vouloir supprimer le message ?')">Supprimer le commentaire</a>
-				</div>
+					<div class="container bgcolor">
+						<p>ID du commentaire: <?php echo $commentaire["comment_ID"]; ?></p>
+						<p>ID du post: <?php echo $commentaire["comment_post_ID"]; ?></p>
+						<p>ID auteur: <?php echo $commentaire["comment_author"]; ?></p>	
+						<p>Date: <?php echo $commentaire["comment_date"]; ?></p>
+						<p>Contenu: <?php echo $commentaire["comment_content"]; ?></p>
+						<a href="?module=commentaire&action=supprimer_commentaire&page=<?= $page ?>&id=<?= $commentaire['comment_ID']; ?>" onclick=" return confirm('Etes-vous certain de vous certain de vouloir supprimer le message ?')">Supprimer le commentaire</a>
+					</div>
 			<?php } ?>
 		<?php } ?>
